requirements.txt
streamlit
spacy
langchain
langchain-anthropic
langchain-openai
python-dotenv
PyPDF2
pdfplumber
textstat
pandas
python-docx
plotly
watchdog
reportlab
langdetect
pymongo
anthropic
google-generativeai
streamlit-option-menu
streamlit-extras
